+++ lib/src/game/padlocks/hint.dart
@@ -9,13 +9,9 @@ class PadlockHint {
   /// The hint text.
   final String text;
 
-  /// The minimum try count before being able to read this hint.
-  final int minimumTriesBeforeShowing;
-
   /// Creates a new [PadlockHint] instance.
   const PadlockHint({
     this.title = kDefaultPadlockHintTitle,
     required this.text,
-    required this.minimumTriesBeforeShowing,
   });
 }